Interfacing Model with Graph
- Some links needed to be built between Graph and Model (and more
particularly Agents)
- For example, when importing an Agent, not only the corresponding node
should be added to the Graph, but moreover the task corresponding to
the Agent should be added to the corresponding AgentGroup's job : the
graph could not handle it.
- Moreover, we don't want to directly integrate Model, Scheduler and
Runtime into the DistributedGraph API
- The adopted solution corresponds to the definition of "callback"
functions on the graph, that are called when nodes are inserted or
deleted.
- Such a feature is also really powerful to allocate / delete
resources contained in graph Nodes, for example Agent* pointers in our
case : when the node is "erased", deleting the contained Agent* pointer
is safe.