Graph / Model integration : fixes data updates
- Agents are now wrapped in an AgentPtrWrapper, that implements
unique_ptr like semantics.
- In consequence, graph::buildNode(const T&) needed to be removed...
- Instead, data are always "moved" into the graph.
- This also fixes data updates, in GhostMode as in HardSyncMode :
simply overriding the AgentPtr with a new AgentPtr (containing a new
dynamically allocated Agent) caused memory leaks, and overrides the Task
associated to the original Agent (that is associated once to the agent,
when it is 'setLocal', not each time the Agent is updated).
- to solve this, updates are "moved" to the original AgentPtrWrapper,
and the original Task is properly rebind.