ReaderAgent HardSyncMode Model MPI integration
- a lot of bugs has been fixed
- AgentPtrWrapper is now CopyAssignable, because it was a requirement to
work in HardSyncMode. However, Copy and Move should currently be used
with extra caution, due to Task / Node bindings to Agents...
- When importing an existing node into the graph, weight and data of the
local node is updated according to the imported node
- An important deadlock has been detected in HardSyncMode, when
synchronizing the graph. To solve it, the TerminationAlgorithm should be
adapted. As a quick fix, SyncLinker synchronization has been
TEMPORARILY disabled, but this will be quickly fixed in a next commit.