Improves location management on Arc import
- Finally, there is no need to update locations with the LocationManager
in the DistributedGraph::synchronize function.
- this method was inefficient, and could not be applied to the HardSync
mode, since the location of the potentially created DISTANT nodes when an Arc is
imported must be updated directly, because some Task might access the new
DISTANT node before DistributedGraph::synchronize is called.
- In consequence, src and tgt locations are now bundled in the Arc
serialization, and the src and tgt locations are set directly in the Arc
deserialization process.
- This method is more efficient, since is does not requires extra
communication to update locations, and works in GhostMode as in
HardSyncMode.