:sparkles: Introduces a light json concept to optimize edge transmission
This is a work in progress, that will be improved in next commits.

- DistributedEdges are now transmitted without serializing source and
target nodes data.
- However, this requires Agents to be default constructible, what might
introduce incompatibilities
- Finally, the DistributedMoveAlgorithm implementation has been
improved. More data synchronizations are required, but global
performances might still be improved since this is only due to the
DistributedEdge serialization optimization. Moreover, read operations
are performed on Cells when building perception and mobility fields,
still due to the fact that nodes data is not systematically imported
with new edges. But this is theoretically safer, since this allows
ranges to depend on dynamic cell parameters.