:bug: Solves a FileOutput bug
When a FileOuput was created on all processes in std::ios::out mode,
the file was opened on all the processes and so the file content was
deleted on all processes, even if only one process was exepected to
write to the file, what produces unexpected results, especially when
writing the CSV header with a DistributedCsvOutput.

This has been solved by opening the file only when FileOuput::get() is
called for the first time, what should not happen on processes that are
not expected to write to the file, so they won't unexpectedly open the
file.